+++ src/schedule/signal.h
@@ -13,14 +13,14 @@ public:
 
 	signal_t();
 	~signal_t();
-	void notify(size_t data);
+	void notify(int data);
 	void wait(process_t *proc);
 	
 private:
 
 	std::mutex mut;
 	std::queue<process_t*> proc;
-	std::queue<size_t> que;
+	std::queue<int> que;
 };
 
 namespace signal_id {
@@ -36,7 +36,7 @@ void destroy_signal ();
  * @return : 0 ok
  *          -1 not valid signal id
  */ 
-int send_signal(int signal_id, size_t data);
+int send_signal(int signal_id, int data);
 int wait_signal(int signal_id, process_t *proc);
 
 int register_signal();
